BinanceApi/requirements.txt
Flask==2.3.2
plotly==5.15.0
pycoingecko==3.1.0
python-binance==1.0.17
pandas==2.0.2
openpyxl==3.1.2